backend/requirements/base.txt
# ─── Django core ──────────────────────────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
django-environ==0.11.2
django-extensions==3.2.3

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ─── Caching & Async ──────────────────────────────────────────────────────────
redis==5.0.4
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
flower==2.0.1

# ─── Security ─────────────────────────────────────────────────────────────────
cryptography==42.0.5
django-ratelimit==4.1.0
django-axes==6.4.0
argon2-cffi==23.1.0

# ─── File handling ────────────────────────────────────────────────────────────
Pillow==10.3.0
python-magic==0.4.27
PyMuPDF==1.24.2          # PDF processing (fitz)
pytesseract==0.3.10       # OCR
openpyxl==3.1.2           # Excel import/export
pandas==2.2.2             # Data processing
xlrd==2.0.1

# ─── PDF generation ───────────────────────────────────────────────────────────
reportlab==4.1.0
WeasyPrint==61.2
Jinja2==3.1.4

# ─── Storage ──────────────────────────────────────────────────────────────────
django-storages==1.14.3
boto3==1.34.80            # S3 / MinIO

# ─── API & Serialization ──────────────────────────────────────────────────────
drf-spectacular==0.27.2   # OpenAPI / Swagger docs
djangorestframework-camel-case==1.4.2

# ─── Multi-tenancy ────────────────────────────────────────────────────────────
django-tenant-schemas==1.12.0

# ─── Utils ────────────────────────────────────────────────────────────────────
python-slugify==8.0.4
phonenumbers==8.13.36
python-dateutil==2.9.0
humanize==4.9.0
